requirements.txt


streamlit
pandas
sqlalchemy
pymysql
python-dotenv
pymongo
langchain
langchain-openai
streamlit-oauth
matplotlib